@@ -21,26 +21,26 @@
     ]
   },
   "dependencies": {
-    "@formspree/react": "^2.2.4",
+    "@formspree/react": "^2.3.0",
     "@openlab/vercel-netlify-cms-github": "^1.1.1",
     "algoliasearch": "^4.12.1",
     "babel-plugin-prismjs": "^2.1.0",
     "date-fns": "^2.28.0",
-    "eslint": "^7.26.0",
+    "eslint": "^8.19.0",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-react": "^7.23.2",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.30.1",
     "gray-matter": "^4.0.3",
-    "immer": "^9.0.14",
-    "next": "^12.1.6",
+    "immer": "^9.0.15",
+    "next": "^12.2.2",
     "next-pwa": "^5.2.21",
-    "next-seo": "^4.26.0",
-    "next-slicezone": "^0.1.3",
-    "nextjs-progressbar": "^0.0.13",
+    "next-seo": "^5.4.0",
+    "next-slicezone": "^0.2.6",
+    "nextjs-progressbar": "^0.0.14",
     "prismjs": "^1.26.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-markdown": "6.0.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-markdown": "8.0.3",
     "react-masonry-css": "^1.0.16",
     "remark": "13.0.0",
     "remark-autolink-headings": "6.0.1",
@@ -50,23 +50,23 @@
     "styled-components": "^5.3.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.14.2",
-    "@babel/preset-typescript": "^7.13.0",
-    "@types/node": "^17.0.41",
-    "@types/react": "^18.0.12",
-    "@types/react-dom": "^18.0.5",
+    "@babel/core": "^7.18.6",
+    "@babel/preset-typescript": "^7.18.6",
+    "@types/node": "^18.0.3",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
     "@types/styled-components": "^5.1.25",
     "babel-loader": "^8.2.2",
-    "babel-plugin-styled-components": "^1.12.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "babel-plugin-styled-components": "^2.0.7",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "file-loader": "^6.2.0",
-    "globby": "11.0.4",
-    "husky": "^6.0.0",
-    "lint-staged": "^11.0.0",
-    "plop": "^2.7.4",
-    "prettier": "2.3.0",
-    "sitemap": "7.0.0",
-    "typescript": "^4.7.3",
+    "globby": "13.1.2",
+    "husky": "^8.0.1",
+    "lint-staged": "^13.0.3",
+    "plop": "^3.1.1",
+    "prettier": "2.7.1",
+    "sitemap": "7.1.1",
+    "typescript": "^4.7.4",
     "url-loader": "^4.1.1"
   }
 }
